[FrontendTool] Split out swiftdeps generation into its own file.
No intended functionality change.
